Refactor QR component, moving logics from header to component
